requirements.txt
watchdog==6.0.0
Pillow==11.3.0